fix mpi_array_type creation
I don't know if my current fix is the best option, but it seems to be a
working solution to the problem.
I use FFTW to figure out how to distribute the fields between CPUS.
This yields nice results for numbers of CPUs that are powers of 2, and
fields that have dimensions which are powers of 2 (or multiples of large
enough powers of 2).
However, it can keep one or more CPUs empty in case it decides that's
faster (and it probably is).
MPI is stupid, and when I try to call MPI_Type_create_subarray with
subsizes that are 0 it fails telling me that I'm an idiot.
Therefore what I do right now is to keep track of which processes
actually have data, and then only create the subarray for those.

I still need to check whether the results are now sane though.